+++ SpatialGDK/Source/SpatialGDK/Public/Interop/ActorSystem.h
@@ -4,9 +4,9 @@
 #include "ClaimPartitionHandler.h"
 #include "Schema/SpawnData.h"
 #include "Schema/UnrealMetadata.h"
-#include "SpatialConstants.h"
 #include "Utils/RepDataUtils.h"
 
+#include "Interop/ClientNetLoadActorHelper.h"
 #include "Interop/CreateEntityHandler.h"
 
 DECLARE_LOG_CATEGORY_EXTERN(LogActorSystem, Log, All);
@@ -22,7 +22,6 @@ struct FClassInfo;
 class USpatialNetDriver;
 
 class SpatialActorChannel;
-class USpatialNetDriver;
 
 using FChannelsToUpdatePosition =
 	TSet<TWeakObjectPtr<USpatialActorChannel>, TWeakObjectPtrKeyFuncs<TWeakObjectPtr<USpatialActorChannel>, false>>;
@@ -61,7 +60,7 @@ public:
 
 	// Updates
 	void SendComponentUpdates(UObject* Object, const FClassInfo& Info, USpatialActorChannel* Channel, const FRepChangeState* RepChanges,
-							  const FHandoverChangeState* HandoverChanges, uint32& OutBytesWritten);
+							  uint32& OutBytesWritten);
 	void SendActorTornOffUpdate(Worker_EntityId EntityId, Worker_ComponentId ComponentId) const;
 	void ProcessPositionUpdates();
 	void RegisterChannelForPositionUpdate(USpatialActorChannel* Channel);
@@ -79,6 +78,8 @@ public:
 
 	static Worker_ComponentData CreateLevelComponentData(const AActor& Actor, const UWorld& NetDriverWorld,
 														 const USpatialClassInfoManager& ClassInfoManager);
+
+	void DestroySubObject(const Worker_EntityId EntityId, UObject& Object, const FUnrealObjectRef& ObjectRef) const;
 
 private:
 	// Helper struct to manage FSpatialObjectRepState update cycle.
@@ -134,7 +135,6 @@ private:
 	void ApplyComponentData(USpatialActorChannel& Channel, UObject& TargetObject, const Worker_ComponentId ComponentId,
 							Schema_ComponentData* Data);
 
-	bool IsDynamicSubObject(AActor* Actor, uint32 SubObjectOffset);
 	void ResolveIncomingOperations(UObject* Object, const FUnrealObjectRef& ObjectRef);
 	void ResolveObjectReferences(FRepLayout& RepLayout, UObject* ReplicatedObject, FSpatialObjectRepState& RepState,
 								 FObjectReferencesMap& ObjectReferencesMap, uint8* RESTRICT StoredData, uint8* RESTRICT Data,
@@ -143,7 +143,7 @@ private:
 	// Component update
 	USpatialActorChannel* GetOrRecreateChannelForDormantActor(AActor* Actor, Worker_EntityId EntityID) const;
 	void ApplyComponentUpdate(Worker_ComponentId ComponentId, Schema_ComponentUpdate* ComponentUpdate, UObject& TargetObject,
-							  USpatialActorChannel& Channel, bool bIsHandover);
+							  USpatialActorChannel& Channel);
 
 	// Entity add
 	void ReceiveActor(Worker_EntityId EntityId);
@@ -178,6 +178,7 @@ private:
 
 	USpatialNetDriver* NetDriver;
 	SpatialEventTracer* EventTracer;
+	FClientNetLoadActorHelper ClientNetLoadActorHelper;
 
 	CreateEntityHandler CreateEntityHandler;
 	ClaimPartitionHandler ClaimPartitionHandler;