+++ SpatialGDK/Source/SpatialGDKEditor/Private/SchemaGenerator/SpatialGDKEditorSchemaGenerator.cpp
@@ -189,33 +189,6 @@ void CheckIdentifierNameValidity(TSharedPtr<FUnrealType> TypeInfo, bool& bOutSuc
 		}
 	}
 
-	// Check Handover data.
-	FCmdHandlePropertyMap HandoverData = GetFlatHandoverData(TypeInfo);
-	TMap<FString, TSharedPtr<FUnrealProperty>> SchemaHandoverDataNames;
-	for (auto& Prop : HandoverData)
-	{
-		FString NextSchemaHandoverDataName = SchemaFieldName(Prop.Value);
-
-		if (!CheckSchemaNameValidity(NextSchemaHandoverDataName, Prop.Value->Property->GetPathName(), TEXT("Handover property")))
-		{
-			bOutSuccess = false;
-		}
-
-		if (TSharedPtr<FUnrealProperty>* ExistingHandoverData = SchemaHandoverDataNames.Find(NextSchemaHandoverDataName))
-		{
-			UE_LOG(LogSpatialGDKSchemaGenerator, Error,
-				   TEXT("Handover data name collision after removing non-alphanumeric characters, schema not generated. Name '%s' collides "
-						"for '%s' and '%s'"),
-				   *NextSchemaHandoverDataName, *ExistingHandoverData->Get()->Property->GetPathName(),
-				   *Prop.Value->Property->GetPathName());
-			bOutSuccess = false;
-		}
-		else
-		{
-			SchemaHandoverDataNames.Add(NextSchemaHandoverDataName, Prop.Value);
-		}
-	}
-
 	// Check subobject name validity.
 	FSubobjects Subobjects = GetAllSubobjects(TypeInfo);
 	TMap<FString, TSharedPtr<FUnrealType>> SchemaSubobjectNames;
@@ -550,14 +523,16 @@ TMap<Worker_ComponentId, FString> CreateComponentIdToClassPathMap()
 
 FString GetComponentSetNameBySchemaType(ESchemaComponentType SchemaType)
 {
+	static_assert(SCHEMA_Count == 4, "Unexpected number of Schema type components, please check the enclosing function is still correct.");
+
 	switch (SchemaType)
 	{
 	case SCHEMA_Data:
 		return SpatialConstants::DATA_COMPONENT_SET_NAME;
 	case SCHEMA_OwnerOnly:
 		return SpatialConstants::OWNER_ONLY_COMPONENT_SET_NAME;
-	case SCHEMA_Handover:
-		return SpatialConstants::HANDOVER_COMPONENT_SET_NAME;
+	case SCHEMA_ServerOnly:
+		return SpatialConstants::SERVER_ONLY_COMPONENT_SET_NAME;
 	case SCHEMA_InitialOnly:
 		return SpatialConstants::INITIAL_ONLY_COMPONENT_SET_NAME;
 	default:
@@ -571,13 +546,15 @@ FString GetComponentSetNameBySchemaType(ESchemaComponentType SchemaType)
 
 Worker_ComponentId GetComponentSetIdBySchemaType(ESchemaComponentType SchemaType)
 {
+	static_assert(SCHEMA_Count == 4, "Unexpected number of Schema type components, please check the enclosing function is still correct.");
+
 	switch (SchemaType)
 	{
 	case SCHEMA_Data:
 		return SpatialConstants::DATA_COMPONENT_SET_ID;
 	case SCHEMA_OwnerOnly:
 		return SpatialConstants::OWNER_ONLY_COMPONENT_SET_ID;
-	case SCHEMA_Handover:
+	case SCHEMA_ServerOnly:
 		return SpatialConstants::HANDOVER_COMPONENT_SET_ID;
 	case SCHEMA_InitialOnly:
 		return SpatialConstants::INITIAL_ONLY_COMPONENT_SET_ID;
@@ -659,23 +636,8 @@ void WriteServerAuthorityComponentSet(const USchemaDatabase* SchemaDatabase, con
 				const Worker_ComponentId ComponentId = GeneratedActorClass.Value.SchemaComponents[SchemaType];
 				if (ComponentId != 0)
 				{
-					switch (SchemaType)
-					{
-					case SCHEMA_Data:
-						Writer.Printf("unreal.generated.{0}.{1},", ActorClassName.ToLower(), ActorClassName);
-						break;
-					case SCHEMA_OwnerOnly:
-						Writer.Printf("unreal.generated.{0}.{1}OwnerOnly,", ActorClassName.ToLower(), ActorClassName);
-						break;
-					case SCHEMA_Handover:
-						Writer.Printf("unreal.generated.{0}.{1}Handover,", ActorClassName.ToLower(), ActorClassName);
-						break;
-					case SCHEMA_InitialOnly:
-						Writer.Printf("unreal.generated.{0}.{1}InitialOnly,", ActorClassName.ToLower(), ActorClassName);
-						break;
-					default:
-						break;
-					}
+					Writer.Printf("unreal.generated.{0}.{1}{2},", ActorClassName.ToLower(), ActorClassName,
+								  GetReplicatedPropertyGroupName(SchemaComponentTypeToPropertyGroup(SchemaType)));
 				}
 			});
 
@@ -687,23 +649,8 @@ void WriteServerAuthorityComponentSet(const USchemaDatabase* SchemaDatabase, con
 					const Worker_ComponentId& ComponentId = ActorSubObjectData.Value.SchemaComponents[SchemaType];
 					if (ComponentId != 0)
 					{
-						switch (SchemaType)
-						{
-						case SCHEMA_Data:
-							Writer.Printf("unreal.generated.{0}.subobjects.{1},", ActorClassName.ToLower(), ActorSubObjectName);
-							break;
-						case SCHEMA_OwnerOnly:
-							Writer.Printf("unreal.generated.{0}.subobjects.{1}OwnerOnly,", ActorClassName.ToLower(), ActorSubObjectName);
-							break;
-						case SCHEMA_Handover:
-							Writer.Printf("unreal.generated.{0}.subobjects.{1}Handover,", ActorClassName.ToLower(), ActorSubObjectName);
-							break;
-						case SCHEMA_InitialOnly:
-							Writer.Printf("unreal.generated.{0}.subobjects.{1}InitialOnly,", ActorClassName.ToLower(), ActorSubObjectName);
-							break;
-						default:
-							break;
-						}
+						Writer.Printf("unreal.generated.{0}.subobjects.{1}{2},", ActorClassName.ToLower(), ActorSubObjectName,
+									  GetReplicatedPropertyGroupName(SchemaComponentTypeToPropertyGroup(SchemaType)));
 					}
 				});
 			}
@@ -722,23 +669,8 @@ void WriteServerAuthorityComponentSet(const USchemaDatabase* SchemaDatabase, con
 					const Worker_ComponentId& ComponentId = SubObjectSchemaData.SchemaComponents[SchemaType];
 					if (ComponentId != 0)
 					{
-						switch (SchemaType)
-						{
-						case SCHEMA_Data:
-							Writer.Printf("unreal.generated.{0}Dynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-							break;
-						case SCHEMA_OwnerOnly:
-							Writer.Printf("unreal.generated.{0}OwnerOnlyDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-							break;
-						case SCHEMA_Handover:
-							Writer.Printf("unreal.generated.{0}HandoverDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-							break;
-						case SCHEMA_InitialOnly:
-							Writer.Printf("unreal.generated.{0}InitialOnlyDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-							break;
-						default:
-							break;
-						}
+						Writer.Printf("unreal.generated.{0}{1}Dynamic{2},", SubObjectClassName,
+									  GetReplicatedPropertyGroupName(SchemaComponentTypeToPropertyGroup(SchemaType)), SubObjectNumber + 1);
 					}
 				});
 			}
@@ -868,6 +800,8 @@ void WriteComponentSetBySchemaType(const USchemaDatabase* SchemaDatabase, ESchem
 	Writer.Printf("id = {0};", GetComponentSetIdBySchemaType(SchemaType));
 	Writer.Printf("components = [").Indent();
 
+	FString SchemaTypeString = GetReplicatedPropertyGroupName(SchemaComponentTypeToPropertyGroup(SchemaType));
+
 	// Write all components.
 	{
 		for (const auto& GeneratedActorClass : SchemaDatabase->ActorClassPathToSchema)
@@ -876,23 +810,7 @@ void WriteComponentSetBySchemaType(const USchemaDatabase* SchemaDatabase, ESchem
 			const FString& ActorClassName = UnrealNameToSchemaComponentName(GeneratedActorClass.Value.GeneratedSchemaName);
 			if (GeneratedActorClass.Value.SchemaComponents[SchemaType] != 0)
 			{
-				switch (SchemaType)
-				{
-				case SCHEMA_Data:
-					Writer.Printf("unreal.generated.{0}.{1},", ActorClassName.ToLower(), ActorClassName);
-					break;
-				case SCHEMA_OwnerOnly:
-					Writer.Printf("unreal.generated.{0}.{1}OwnerOnly,", ActorClassName.ToLower(), ActorClassName);
-					break;
-				case SCHEMA_Handover:
-					Writer.Printf("unreal.generated.{0}.{1}Handover,", ActorClassName.ToLower(), ActorClassName);
-					break;
-				case SCHEMA_InitialOnly:
-					Writer.Printf("unreal.generated.{0}.{1}InitialOnly,", ActorClassName.ToLower(), ActorClassName);
-					break;
-				default:
-					break;
-				}
+				Writer.Printf("unreal.generated.{0}.{1}{2},", ActorClassName.ToLower(), ActorClassName, SchemaTypeString);
 			}
 			// Actor static subobjects.
 			for (const auto& ActorSubObjectData : GeneratedActorClass.Value.SubobjectData)
@@ -900,23 +818,8 @@ void WriteComponentSetBySchemaType(const USchemaDatabase* SchemaDatabase, ESchem
 				const FString ActorSubObjectName = UnrealNameToSchemaComponentName(ActorSubObjectData.Value.Name.ToString());
 				if (ActorSubObjectData.Value.SchemaComponents[SchemaType] != 0)
 				{
-					switch (SchemaType)
-					{
-					case SCHEMA_Data:
-						Writer.Printf("unreal.generated.{0}.subobjects.{1},", ActorClassName.ToLower(), ActorSubObjectName);
-						break;
-					case SCHEMA_OwnerOnly:
-						Writer.Printf("unreal.generated.{0}.subobjects.{1}OwnerOnly,", ActorClassName.ToLower(), ActorSubObjectName);
-						break;
-					case SCHEMA_Handover:
-						Writer.Printf("unreal.generated.{0}.subobjects.{1}Handover,", ActorClassName.ToLower(), ActorSubObjectName);
-						break;
-					case SCHEMA_InitialOnly:
-						Writer.Printf("unreal.generated.{0}.subobjects.{1}InitialOnly,", ActorClassName.ToLower(), ActorSubObjectName);
-						break;
-					default:
-						break;
-					}
+					Writer.Printf("unreal.generated.{0}.subobjects.{1}{2},", ActorClassName.ToLower(), ActorSubObjectName,
+								  SchemaTypeString);
 				}
 			}
 		}
@@ -931,23 +834,7 @@ void WriteComponentSetBySchemaType(const USchemaDatabase* SchemaDatabase, ESchem
 					GeneratedSubObjectClass.Value.DynamicSubobjectComponents[SubObjectNumber];
 				if (SubObjectSchemaData.SchemaComponents[SchemaType] != 0)
 				{
-					switch (SchemaType)
-					{
-					case SCHEMA_Data:
-						Writer.Printf("unreal.generated.{0}Dynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-						break;
-					case SCHEMA_OwnerOnly:
-						Writer.Printf("unreal.generated.{0}OwnerOnlyDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-						break;
-					case SCHEMA_Handover:
-						Writer.Printf("unreal.generated.{0}HandoverDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-						break;
-					case SCHEMA_InitialOnly:
-						Writer.Printf("unreal.generated.{0}InitialOnlyDynamic{1},", SubObjectClassName, SubObjectNumber + 1);
-						break;
-					default:
-						break;
-					}
+					Writer.Printf("unreal.generated.{0}{1}Dynamic{2},", SubObjectClassName, SchemaTypeString, SubObjectNumber + 1);
 				}
 			}
 		}
@@ -974,8 +861,9 @@ void WriteComponentSetFiles(const USchemaDatabase* SchemaDatabase, FString Schem
 	WriteRoutingWorkerAuthorityComponentSet(SchemaOutputPath);
 	WriteComponentSetBySchemaType(SchemaDatabase, SCHEMA_Data, SchemaOutputPath);
 	WriteComponentSetBySchemaType(SchemaDatabase, SCHEMA_OwnerOnly, SchemaOutputPath);
-	WriteComponentSetBySchemaType(SchemaDatabase, SCHEMA_Handover, SchemaOutputPath);
+	WriteComponentSetBySchemaType(SchemaDatabase, SCHEMA_ServerOnly, SchemaOutputPath);
 	WriteComponentSetBySchemaType(SchemaDatabase, SCHEMA_InitialOnly, SchemaOutputPath);
+	static_assert(SCHEMA_Count == 4, "Unexpected number of Schema type components, please check the enclosing function is still correct.");
 }
 
 USchemaDatabase* InitialiseSchemaDatabase(const FString& PackagePath)